discord_bot/requirements.txt
supabase==2.10.0
httpx>=0.26,<0.28
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
shortuuid==1.0.11
discord.py==2.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 